backend/requirements.txt
Django>=5.0.0,<6.0.0
djangorestframework>=3.14.0,<4.0.0
djangorestframework-simplejwt>=5.3.0,<6.0.0
django-cors-headers>=4.3.0,<5.0.0
channels[daphne]>=4.0.0,<5.0.0
channels-redis>=4.2.0,<5.0.0
psycopg2-binary>=2.9.0,<3.0.0
python-decouple>=3.8,<4.0
django-filter>=24.0.0,<25.0.0
gunicorn>=21.2.0,<23.0.0
whitenoise>=6.6.0,<7.0.0
